requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pyjwt
slowapi
pydantic
google-generativeai

